+++ tcf_website/static/qa/qa_dashboard.css
@@ -609,10 +609,6 @@
   background: var(--secondary-color);
 }
 
-.empty-state {
-  text-align: center;
-}
-
 /* Responsive Design */
 @media (max-width: 768px) {
   .qa-container {
