+++ tcf_website/templates/qa/qa_dashboard.html
@@ -11,7 +11,7 @@
 {% block content %}
     <div class="qa-container">
         <div class="qa-sidebar">
-            <!-- TODO: Add functionality to post button and searching-->
+            <!-- TODO: Add functionality to post/filter button and searching-->
             <div class="qa-header">
                 <button class="btn-new-post">
                     <i class="fa fa-plus"></i> New Post
@@ -21,6 +21,15 @@
                     <input type="text" class="search-input" placeholder="Search posts...">
                 </div>
             </div>
+
+            <div class="qa-filters">
+                <button class="filter-btn active">
+                    <i class="fa fa-list"></i> Filter
+                </button>
+                <div class="dropdown">
+                </div>
+            </div>
+
             <!-- Sidebar Post List -->
             <div class="posts-list">
                 {% if questions %}
@@ -43,11 +52,104 @@
             </div>
         </div>
 
-        <!-- Post Content -->
+        <!-- Right Side Content -->
          <div class="qa-content">
+            {% if active_question %}
+            <div class="content-nav">
+                <div class="post-info">
+                    <h3>{{ active_question.course.code }}</h3>
+                </div>
+                <button class="nav-btn">
+                    <i class="fa fa-ellipsis-v"></i>
+                </button>
+            </div>
 
+            <!-- Post Content -->
+             <div class="post-content">
+                <div class="main-question">
+                    <h2 class="post-content-title"> {{ active_question.title }}</h2>
+
+                    <div class="post-meta">
+                        <span>Asked {{ active_question.created|timesince }} ago</span>
+                    </div>
+
+                    <div class="post-body">
+                        <p> {{ active_question.text }}</p>
+                    </div>
+
+                    <!-- TODO: Increment count for upvote. Also adding edit button? -->
+                    <div class="post-actions">
+                        <button class="action-btn">
+                            <i class="fa fa-thumbs-up"></i>
+                            {{ active_question.sum_q_votes|default:0 }}
+                        </button>
+                    </div>
+                </div>
+
+                <!-- Responses -->
+                <div class="comments-section">
+                    <h3 class="comments-header">Responses</h3>
+                </div>
+             </div>
+            {% else %}
+                <p class="empty-state">No responses yet.</p>
+            {% endif %}
          </div>
     </div> 
+
+
+    <!-- New Post Modal -->
+    <div id="newPostModal" class="modal-overlay" style="display:none">
+        <div class="modal-container">
+            <div class="modal-header">
+                <h2>Create New Post</h2>
+                <button class="modal-close" onclick="closeModal()">
+                    <i class="fa fa-times"></i>
+                </button>
+            </div>
+
+            <form method="post" action="{% url 'create_question' %}">
+                {% csrf_token %}
+
+                <div class="modal-body">
+                    <div class="form-group">
+                        <label for="course">Course</label>
+                        <select name="course" required>
+                            <!-- Doing this causes a lot of lag -->
+                            <!-- {% for course in course %}
+                                <option value="{{ course.id }}">
+                                    {{ course.code }}
+                                </option>
+                            {% endfor %} -->
+                            <option>2100</option>
+                        </select>
+                    </div>
+
+                    <div class="form-group">
+                        <label for="text">Questions</label>
+                        <textarea
+                            name="text"
+                            placeholder="Ask your question..."
+                            required
+                        ></textarea>
+                    </div>
+                </div>
+
+                <div class="modal-footer">
+                    <button type="button" class="btn-cancel" onclick="closeModal()">
+                        Cancel
+                    </button>
+                    <button type="submit" class="btn-submit">
+                        Create Post
+                    </button>
+                </div>
+            </form>
+        </div>
+    </div>
         
      
+{% endblock %}
+
+{% block js %}
+    <script src="{% static 'qa/qa_dashboard.js' %}"></script>
 {% endblock %}
