+++ tcf_website/templates/qa/qa_dashboard_hard.html
@@ -258,5 +258,5 @@
 {% endblock %}
     
 {% block js %}
-    <script src="{% static 'qa/qa_dashboard.js' %}"></script>
+    <script src="{% static 'qa/qa_dashboard_test.js' %}"></script>
 {% endblock %}
