+++ tcf_website/urls.py
@@ -100,6 +100,7 @@ urlpatterns = [
     path("qa/", views.qa_dashboard, name="qa"),
     path("qaTest/", views.qa_dashboard_hard, name="qaTest"),
     path("forum/", views.forum.forum_dashboard, name="forum_dashboard"),
+    path("qa/create/", views.qa.create_question, name="create_question"),
     # Post detail (AJAX)
     path(
         "forum/post/<int:post_id>/",
